requirements.txt
fastapi
uvicorn
python-socketio
aiortc
aiohttp
cryptography
netifaces
deep-translator
pandas
openpyxl
python-multipart
